fix score view

+++ app/src/main/java/com/cretihoy/wordminded/data/db/entity/User.kt
@@ -13,5 +13,5 @@ data class User(
     val name: String,
 
     @ColumnInfo
-    val score: Long
+    val score: Int
 )
